+++ .eslintignore
@@ -1,4 +1,4 @@
 src/index.js
-src/App.test.js
+src/tests/*.test.js
 src/serviceWorker.js
 src/services/movieData.js
